Revert "Split tests in multiple files for parallel execution."
